// src/modules/battle/constants/battle.enums.ts
export enum BattleModeEnum {
  CLASSIC = 'classic',
  RANKED = 'ranked',
  TOURNAMENT = 'tournament',
  DUOS = 'duos',
}

export enum BattleStatusEnum {
  IN_PROGRESS = 'in_progress',
  COMPLETED = 'completed',
  CANCELLED = 'cancelled',
}

export enum BattleOpponentTypeEnum {
  PVP = 'pvp',
  BOT = 'bot',
}

export enum BattleRegionEnum {
  IR = 'IR',
  GLOBAL = 'GLOBAL',
}

export enum BattleWinnerEnum {
  PLAYER1 = 'player1',
  PLAYER2 = 'player2',
  DRAW = 'draw',
}

export enum BattleStatCategoryEnum {
  SPEED = 'speed',
  POWER = 'power',
  SKILL = 'skill',
  ATTACK = 'attack',
  DEFEND = 'defend',
}